advances presentation of weighted means for smaples per discussions with @NicoleRayner
