+++ squidApp/src/main/java/org/cirdles/squid/gui/dateInterpretations/plots/squid/WeightedMeanPlot.java
@@ -43,7 +43,7 @@ import javafx.scene.text.Text;
 import org.cirdles.squid.gui.dataViews.AbstractDataView;
 import org.cirdles.squid.gui.dataViews.TicGeneratorForAxes;
 import org.cirdles.squid.gui.dateInterpretations.plots.PlotDisplayInterface;
-import org.cirdles.squid.gui.dateInterpretations.plots.PlotsController;
+import org.cirdles.squid.gui.dateInterpretations.plots.plotControllers.PlotsController;
 import org.cirdles.squid.shrimp.ShrimpFractionExpressionInterface;
 import org.cirdles.squid.tasks.expressions.spots.SpotSummaryDetails;
 import static org.cirdles.squid.utilities.conversionUtilities.RoundingUtilities.squid3RoundedToSize;
@@ -58,13 +58,14 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
     private SpotSummaryDetails spotSummaryDetails;
     private List<ShrimpFractionExpressionInterface> shrimpFractions;
     private List<ShrimpFractionExpressionInterface> storedShrimpFractions;
-    private List<Double> ages;
-    private List<Double> ageTwoSigma;
+    private List<Double> agesOrValues;
+    private List<Double> agesOrValuesTwoSigma;
     private List<Double> hours;
     private double[] weightedMeanStats;
     private double[] onPeakTwoSigma;
     private boolean[] rejectedIndices;
-    private String ageLookupString;
+    private String ageOrValueLookupString;
+    private boolean adaptToAgeInMA;
     private int countOfIncluded;
 
     private final double referenceMaterialAge;
@@ -91,8 +92,9 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
         this.plotTitle = plotTitle;
         this.spotSummaryDetails = spotSummaryDetails;
         // extract needed values
-        this.ageLookupString = ageLookupString;
-        extractFractionDetails();
+        this.ageOrValueLookupString = ageLookupString;
+        this.adaptToAgeInMA = ageOrValueLookupString.contains("Age");
+        extractSpotDetails();
 
         this.referenceMaterialAge = referenceMaterialAge;
         this.weightedMeanRefreshInterface = weightedMeanRefreshInterface;
@@ -148,7 +150,7 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
         return this.getWidth();
     }
 
-    private boolean extractFractionDetails() {
+    private boolean extractSpotDetails() {
         storedShrimpFractions = spotSummaryDetails.getSelectedSpots();
         shrimpFractions = new ArrayList<>();
 
@@ -172,9 +174,9 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
                     switch (sortFlavor) {
                         case "AGE":
                             valueFromNode1 = fraction1
-                                    .getTaskExpressionsEvaluationsPerSpotByField(ageLookupString)[0][0];
+                                    .getTaskExpressionsEvaluationsPerSpotByField(ageOrValueLookupString)[0][0];
                             valueFromNode2 = fraction2
-                                    .getTaskExpressionsEvaluationsPerSpotByField(ageLookupString)[0][0];
+                                    .getTaskExpressionsEvaluationsPerSpotByField(ageOrValueLookupString)[0][0];
                             break;
                         case "RATIO":
                             double[][] resultsFromNode1
@@ -205,16 +207,29 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
                 countOfIncluded = countOfIncluded + (rejected ? 0 : 1);
             }
 
-            ages = new ArrayList<>();
-            ageTwoSigma = new ArrayList<>();
+            agesOrValues = new ArrayList<>();
+            agesOrValuesTwoSigma = new ArrayList<>();
             hours = new ArrayList<>();
 
             double index = 0;
-            for (ShrimpFractionExpressionInterface sf : shrimpFractions) {
-                ages.add(sf.getTaskExpressionsEvaluationsPerSpotByField(ageLookupString)[0][0]);
-                ageTwoSigma.add(2.0 * sf.getTaskExpressionsEvaluationsPerSpotByField(ageLookupString)[0][1]);
+            for (ShrimpFractionExpressionInterface spot : shrimpFractions) {
+                double[][] results;
+                if (ageOrValueLookupString.startsWith("/", 3)) {
+                    // ratio case
+                    results = Arrays.stream(spot.getIsotopicRatioValuesByStringName(ageOrValueLookupString)).toArray(double[][]::new);
+                } else {
+                    results = spot.getTaskExpressionsEvaluationsPerSpotByField(ageOrValueLookupString);
+                }
+
+                agesOrValues.add(results[0][0]);
+                // handle no uncertainty case
+                if (results[0].length < 2) {
+                    agesOrValuesTwoSigma.add(0.0);
+                } else {
+                    agesOrValuesTwoSigma.add(2.0 * results[0][1]);
+                }
                 if (viewSortOrder == 0) {
-                    hours.add(sf.getHours());
+                    hours.add(spot.getHours());
                 } else {
                     hours.add(index++);
                 }
@@ -310,25 +325,32 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
 
         if (PlotsController.plotTypeSelected.compareTo(PlotsController.PlotTypes.WEIGHTED_MEAN_SAMPLE) == 0) {
             // section for sample wms
-            text.setText("Wtd Mean of " + ageLookupString);
+            text.setText("Wtd Mean of " + ageOrValueLookupString);
             textWidth = (int) text.getLayoutBounds().getWidth();
             g2d.fillText(text.getText(), rightOfText - textWidth, 75);
-            g2d.fillText(squid3RoundedToSize(weightedMeanStats[0] / 1e6, 5) + " Ma", rightOfText + offset, 75);
-
+            if (adaptToAgeInMA) {
+                g2d.fillText(squid3RoundedToSize(weightedMeanStats[0] / 1e6, 5) + " Ma", rightOfText + offset, 75);
+            } else {
+                g2d.fillText(squid3RoundedToSize(weightedMeanStats[0], 5) + "", rightOfText + offset, 75);
+            }
             text.setText("1-sigmaAbs");
             textWidth = (int) text.getLayoutBounds().getWidth();
             g2d.fillText(text.getText(), rightOfText - textWidth, 95);
-            g2d.fillText(squid3RoundedToSize(weightedMeanStats[1] / 1e6, 5) + " Ma", rightOfText + offset, 95);
+            if (adaptToAgeInMA) {
+                g2d.fillText(squid3RoundedToSize(weightedMeanStats[1] / 1e6, 5) + " Ma", rightOfText + offset, 95);
+            } else {
+                g2d.fillText(squid3RoundedToSize(weightedMeanStats[1], 5) + "", rightOfText + offset, 95);
+            }
 
             text.setText("err 68");
             textWidth = (int) text.getLayoutBounds().getWidth();
             g2d.fillText(text.getText(), rightOfText - textWidth, 115);
-            g2d.fillText(squid3RoundedToSize(weightedMeanStats[2] / 1e6, 5) + "", rightOfText + offset, 115);
+            g2d.fillText(squid3RoundedToSize(weightedMeanStats[2] / (adaptToAgeInMA ? 1e6 : 1.0), 5) + "", rightOfText + offset, 115);
 
             text.setText("err 95");
             textWidth = (int) text.getLayoutBounds().getWidth();
             g2d.fillText(text.getText(), rightOfText - textWidth, 135);
-            g2d.fillText(squid3RoundedToSize(weightedMeanStats[3] / 1e6, 5) + "", rightOfText + offset, 135);
+            g2d.fillText(squid3RoundedToSize(weightedMeanStats[3] / (adaptToAgeInMA ? 1e6 : 1.0), 5) + "", rightOfText + offset, 135);
 
             text.setText("MSWD");
             textWidth = (int) text.getLayoutBounds().getWidth();
@@ -347,7 +369,7 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
 
         } else {
 
-            text.setText("Wtd Mean of Ref Mat Pb/" + ((String) (ageLookupString.contains("Th") ? "Th" : "U")) + " calibr.");
+            text.setText("Wtd Mean of Ref Mat Pb/" + ((String) (ageOrValueLookupString.contains("Th") ? "Th" : "U")) + " calibr.");
             textWidth = (int) text.getLayoutBounds().getWidth();
             g2d.fillText(text.getText(), rightOfText - textWidth, 75);
             g2d.fillText(Double.toString(weightedMeanStats[0]), rightOfText + offset, 75);
@@ -466,13 +488,27 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
                             mapX(minX), mapY(ticsY[i].doubleValue()), mapX(maxX), mapY(ticsY[i].doubleValue()));
 
                     // left side
-                    g2d.fillText(ticsY[i].movePointLeft(6).toBigInteger().toString(),//
-                            (float) mapX(minX) - 25f,
-                            (float) mapY(ticsY[i].doubleValue()) + verticalTextShift);
+                    if (adaptToAgeInMA) {
+                        g2d.fillText(ticsY[i].movePointLeft(6).toBigInteger().toString(),//
+                                (float) mapX(minX) - 25f,
+                                (float) mapY(ticsY[i].doubleValue()) + verticalTextShift);
+                    } else {
+                        g2d.fillText(ticsY[i].toString(),//
+                                (float) mapX(minX) - 25f,
+                                (float) mapY(ticsY[i].doubleValue()) + verticalTextShift);
+                    }
+
                     // right side
-                    g2d.fillText(ticsY[i].movePointLeft(6).toBigInteger().toString(),//
-                            (float) mapX(maxX) + 5f,
-                            (float) mapY(ticsY[i].doubleValue()) + verticalTextShift);
+                    if (adaptToAgeInMA) {
+                        g2d.fillText(ticsY[i].movePointLeft(6).toBigInteger().toString(),//
+                                (float) mapX(maxX) + 5f,
+                                (float) mapY(ticsY[i].doubleValue()) + verticalTextShift);
+                    } else {
+                        g2d.fillText(ticsY[i].toString(),//
+                                (float) mapX(maxX) + 5f,
+                                (float) mapY(ticsY[i].doubleValue()) + verticalTextShift);
+                    }
+
                 }
             }
         }
@@ -500,7 +536,7 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
 
         // Y - label
         if (PlotsController.plotTypeSelected.compareTo(PlotsController.PlotTypes.WEIGHTED_MEAN_SAMPLE) == 0) {
-            text.setText("Age (Ma)");
+            text.setText((adaptToAgeInMA ? "Age (Ma)" : ageOrValueLookupString));
         } else {
             text.setText("Ref Mat Age (Ma)");
         }
@@ -596,9 +632,9 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
     @Override
     public void preparePanel() {
 
-        myOnPeakData = ages.stream().mapToDouble(Double::doubleValue).toArray();
+        myOnPeakData = agesOrValues.stream().mapToDouble(Double::doubleValue).toArray();
         myOnPeakNormalizedAquireTimes = hours.stream().mapToDouble(Double::doubleValue).toArray();
-        onPeakTwoSigma = ageTwoSigma.stream().mapToDouble(Double::doubleValue).toArray();
+        onPeakTwoSigma = agesOrValuesTwoSigma.stream().mapToDouble(Double::doubleValue).toArray();
 
         minY = Double.MAX_VALUE;
         maxY = -Double.MAX_VALUE;
@@ -616,7 +652,7 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
         minX -= xMarginStretch;
         maxX += xMarginStretch;
 
-        // Y-axis is ages
+        // Y-axis is agesOrValues
         minY = Double.MAX_VALUE;
         maxY = -Double.MAX_VALUE;
 
@@ -652,7 +688,7 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
         });
         spotContextMenu.getItems().addAll(menuItem1);
 
-        MenuItem menuItem2 = new MenuItem("Toggle Plot Rejected Spots");
+        MenuItem menuItem2 = new MenuItem("Toggle Show Rejected Spots");
         menuItem2.setOnAction((evt) -> {
             doPlotRejectedSpots = !doPlotRejectedSpots;
             refreshPanel(false, false);
@@ -674,7 +710,7 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
 
     @Override
     public Node displayPlotAsNode() {
-        if (extractFractionDetails()) {
+        if (extractSpotDetails()) {
             preparePanel();
             this.repaint();
         }
@@ -701,10 +737,10 @@ public class WeightedMeanPlot extends AbstractDataView implements PlotDisplayInt
     }
 
     /**
-     * @param ageLookupString the ageLookupString to set
+     * @param ageOrValueLookupString the ageOrValueLookupString to set
      */
-    public void setAgeLookupString(String ageLookupString) {
-        this.ageLookupString = ageLookupString;
+    public void setAgeOrValueLookupString(String ageOrValueLookupString) {
+        this.ageOrValueLookupString = ageOrValueLookupString;
     }
 
 }
