+++ squidApp/src/main/resources/org/cirdles/squid/gui/dateInterpretations/plots/plotControllers/Plots.fxml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.RadioButton?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.control.ToggleGroup?>
+<?import javafx.scene.control.ToolBar?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox fx:id="vboxMaster" minHeight="700.0" minWidth="1200.0" stylesheets="@../../css/projectManager.css" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.cirdles.squid.gui.dateInterpretations.plots.plotControllers.PlotsController">
+   <children>
+      <SplitPane dividerPositions="0.2287145242070117" VBox.vgrow="ALWAYS">
+         <items>
+            <ScrollPane fx:id="spotListScrollPane">
+               <content>
+                  <AnchorPane fx:id="spotListAnchorPane" prefHeight="200.0" prefWidth="200.0" />
+               </content>
+            </ScrollPane>
+            <ScrollPane fx:id="plotScrollPane" hbarPolicy="NEVER" vbarPolicy="NEVER">
+               <content>
+                  <VBox fx:id="plotVBox" alignment="BASELINE_LEFT" prefHeight="695.0" prefWidth="914.0">
+                     <children>
+                        <AnchorPane fx:id="plotAndConfigAnchorPane" VBox.vgrow="ALWAYS" />
+                        <ToolBar fx:id="plotToolBar" prefHeight="40.0" style="-fx-background-color: #67ccff;" VBox.vgrow="NEVER" />
+                        <HBox fx:id="controlPanel" alignment="CENTER_LEFT" minHeight="28.0" prefHeight="28.0" spacing="5.0">
+                           <children>
+                              <Label minWidth="83.0" prefHeight="28.0" prefWidth="83.0" text="Correction:" />
+                              <RadioButton fx:id="corr4_RadioButton" minWidth="68.0" mnemonicParsing="false" onAction="#selectedIsotopeIndexAction" prefHeight="18.0" prefWidth="68.0" selected="true" text="4-corr">
+                                 <toggleGroup>
+                                    <ToggleGroup fx:id="correctionToggleGroup" />
+                                 </toggleGroup>
+                              </RadioButton>
+                              <RadioButton fx:id="corr7_RadioButton" minWidth="68.0" mnemonicParsing="false" onAction="#selectedIsotopeIndexAction" prefHeight="18.0" prefWidth="68.0" text="7-corr" toggleGroup="$correctionToggleGroup" />
+                              <RadioButton fx:id="corr8_RadioButton" minWidth="71.0" mnemonicParsing="false" onAction="#selectedIsotopeIndexAction" prefHeight="18.0" prefWidth="71.0" text="8-corr" toggleGroup="$correctionToggleGroup" />
+                              <Label minWidth="49.0" text="Flavor:" />
+                              <RadioButton fx:id="plotFlavorOneRadioButton" mnemonicParsing="false" onAction="#plotChooserAction" selected="true" text="Plot Flavor One">
+                                 <toggleGroup>
+                                    <ToggleGroup fx:id="plotFlavorToggleGroup" />
+                                 </toggleGroup>
+                              </RadioButton>
+                              <RadioButton fx:id="plotFlavorTwoRadioButton" minWidth="116.0" mnemonicParsing="false" onAction="#plotChooserAction" text="PlotFlavor Two" toggleGroup="$plotFlavorToggleGroup" />
+                              <CheckBox fx:id="autoExcludeSpotsCheckBox" graphicTextGap="0.0" minWidth="141.0" mnemonicParsing="false" onAction="#autoExcludeSpotsCheckBoxAction" prefHeight="18.0" prefWidth="141.0" selected="true" text="Auto-reject spots" />
+                           </children>
+                        </HBox>
+                     </children>
+                  </VBox>
+               </content>
+            </ScrollPane>
+         </items>
+      </SplitPane>
+   </children>
+</VBox>
